fixed tests committess helpers
